新增Shushu&Colldier&PingPongMovement

+++ OpenGL_2/Midtern/SceneObject/Background.cpp
@@ -2,7 +2,7 @@
 
 Background::Background(int number, mat4& matModelView, mat4& matProjection, GLuint shaderHandle) {
 	for (int i = 0; i < number; i++) {
-		Star* pStar = new Star(MIN_STAR_SPEED,MIN_STAR_ROTATE_SPEED,false, matModelView, matProjection, shaderHandle);
+		Star* pStar = new Star(MIN_STAR_SPEED, MIN_STAR_ROTATE_SPEED, false, matModelView, matProjection, shaderHandle);
 		pStar->poolParent = &_starPool;
 		_starPool.Init(pStar);
 	}
@@ -26,13 +26,13 @@ void Background::Update(float delta) {
 	}
 }
 
-void Background::GenerateStar(){
+void Background::GenerateStar() {
 	std::vector<Star*>* usingStars = _starPool.GetUsingObjs();
 	//if star at playGround is lower than MAX_NUM get it and set random position and speed for i
 	if ((int)usingStars->size() < MAX_STAR_NUM) {
 		Star* tmp = _starPool.GetPoolObject();
 		if (tmp != NULL) {
-			tmp->Init(vec3(Random::GetRand(WIDTH), Random::GetRand(HEIGHT), 0.0f), Random::GetRand(MAX_STAR_SPEED, MIN_STAR_SPEED), vec3(Random::GetRand(MAX_STAR_SCALE, MIN_STAR_SCALE)));
+			tmp->Init(vec3(Random::GetRand((float)WIDTH), Random::GetRand((float)HEIGHT), 0.0f), Random::GetRand((float)MAX_STAR_SPEED, (float)MIN_STAR_SPEED), vec3(Random::GetRand(MAX_STAR_SCALE, MIN_STAR_SCALE)));
 			tmp->SetColor(vec4(1.0f, 1.0f, 1.0f, Random::GetRand(MAX_STAR_ALPHA, MIN_STAR_ALPHA)));
 			tmp->SetRotation(Random::GetRand(MAX_STAR_ROTATE_SPEED, MIN_STAR_ROTATE_SPEED), Random::GetRandBool());
 		}
